Add slippy map implementation
This adds the initial slippy map implementation, using OSM's tileserver
to download and cache pre-rendered tiles and drawing them with Skia.

This does not yet implement a way to preload maps for offline use,
however once loaded existing maps can be used offline. Geocoding and
routing are not implemented yet.