Add offline geocoding support
This patch implements fully offline geocoding based on geocoder-nlp
using datasets provided by modrana.org. While the geocoding quality is
not on par with online geocoding, it provides good results in reasonable
time without requiring too much storage.

Since geocoder-nlp and libpostal both aren't commonly found in
repositories, we compile both of these statically. This takes a
significant amount of time for libpostal. The dependencies sqlite3,
kyotocabinet, and marisa are all linked to dynamically.

Since GPS support is currently not present, we just pass the current map
center as a reference point for the search.