Switch to dark map tile style
This switches the tileserver from the official OSM server to a dark
theme created using jawg.io. While the number of tile requests are
limited, this should serve well as a dark theme until potentially
setting up a custom tileserver.

Since the dark theme uses colors matching the app theme, this patch also
made a small change to the search button to surround it with a border.
This should help avoid the button background blending in with the map.

This includes the jawg.io auth token in the application, with minor
obfuscation applied. While this could lead to some abuse, the
consequences should be pretty irrelevant considering it is a free
account anyway.