Switch filesystem tile storage to sqlite
Previously tiles were stored on the filesystem in a simple directory
structure. While this was simple, it makes it more difficult to keep
track of the tile's last access times.

The tile storage has been moved to a single sqlite file now, placed at
~/.cache/charon/tiles.sqlite. Besides the tile data itself, this keeps
track of both access and creation time.

This new database allows us to easily delete the oldest entries based on
last access time, which is automatically updated on retrieval (`GET`).
It also makes it simple to determine whether a tile is older than our
max 'ideal' age, at which point we can attempt to refresh it from the
tileserver or keep using it if we're offline.

Closes #3.