Persist search across view transitions
Previously the geocoding search was cleared whenever the search view was
opened, which can be annoying when trying to look at results on the map
without having success on the first attempt.

To allow easily trying different search results and modifying an
existing search with extra details, the query and its results are now
only cleared when the user explicitly does so themselves.