Add reverse geocoding support
This adds a new long-press action in the map view which will start
reverse geocoding in the background and open the search view to list the
geocoding results.

Both online and offline reverse geocoding is supported, however since
Nominatim does not support returning more than a single result for
reverse geocoding, we've switched the online geocoder to Photon.

Closes #9.